+++ opentelemetry-otlp/examples/basic-otlp-http/src/main.rs
@@ -1,34 +1,36 @@
+use log::{info, Level};
 use once_cell::sync::Lazy;
 use opentelemetry::{
     global,
-    metrics::MetricsError,
-    trace::{TraceContextExt, TraceError, Tracer},
+    metrics::{MetricsError, Unit},
+    trace::{TraceContextExt, TraceError, Tracer, TracerProvider as _},
     Key, KeyValue,
 };
-use opentelemetry_appender_tracing::layer::OpenTelemetryTracingBridge;
+use opentelemetry_appender_log::OpenTelemetryLogBridge;
 use opentelemetry_otlp::WithExportConfig;
-use opentelemetry_sdk::logs as sdklogs;
-use opentelemetry_sdk::resource;
 use opentelemetry_sdk::trace as sdktrace;
+use opentelemetry_sdk::{
+    logs::{self as sdklogs, Config},
+    Resource,
+};
 
 use std::error::Error;
-use tracing::info;
-use tracing_subscriber::prelude::*;
+
+static RESOURCE: Lazy<Resource> = Lazy::new(|| {
+    Resource::new(vec![KeyValue::new(
+        opentelemetry_semantic_conventions::resource::SERVICE_NAME,
+        "basic-otlp-example",
+    )])
+});
 
 fn init_logs() -> Result<sdklogs::LoggerProvider, opentelemetry::logs::LogError> {
-    let service_name = env!("CARGO_BIN_NAME");
     opentelemetry_otlp::new_pipeline()
         .logging()
-        .with_log_config(
-            sdklogs::Config::default().with_resource(resource::Resource::new(vec![KeyValue::new(
-                opentelemetry_semantic_conventions::resource::SERVICE_NAME,
-                service_name,
-            )])),
-        )
+        .with_log_config(Config::default().with_resource(RESOURCE.clone()))
         .with_exporter(
             opentelemetry_otlp::new_exporter()
                 .http()
-                .with_endpoint("http://localhost:4318"),
+                .with_endpoint("http://localhost:4318/v1/logs"),
         )
         .install_batch(opentelemetry_sdk::runtime::Tokio)
 }
@@ -41,36 +43,25 @@ fn init_tracer() -> Result<sdktrace::Tracer, TraceError> {
                 .http()
                 .with_endpoint("http://localhost:4318/v1/traces"),
         )
+        .with_trace_config(sdktrace::config().with_resource(RESOURCE.clone()))
         .install_batch(opentelemetry_sdk::runtime::Tokio)
 }
 
-fn init_metrics() -> Result<(), MetricsError> {
-    let export_config = opentelemetry_otlp::ExportConfig {
-        endpoint: "http://localhost:4318/v1/metrics".to_string(),
-        ..opentelemetry_otlp::ExportConfig::default()
-    };
+fn init_metrics() -> Result<opentelemetry_sdk::metrics::SdkMeterProvider, MetricsError> {
     let provider = opentelemetry_otlp::new_pipeline()
         .metrics(opentelemetry_sdk::runtime::Tokio)
         .with_exporter(
             opentelemetry_otlp::new_exporter()
                 .http()
-                .with_export_config(export_config),
+                .with_endpoint("http://localhost:4318/v1/metrics"),
         )
+        .with_resource(RESOURCE.clone())
         .build();
-    provider.map(|_| ())
+    match provider {
+        Ok(provider) => Ok(provider),
+        Err(err) => Err(err),
+    }
 }
-
-const LEMONS_KEY: Key = Key::from_static_str("ex.com/lemons");
-const ANOTHER_KEY: Key = Key::from_static_str("ex.com/another");
-
-static COMMON_ATTRIBUTES: Lazy<[KeyValue; 4]> = Lazy::new(|| {
-    [
-        LEMONS_KEY.i64(10),
-        KeyValue::new("A", "1"),
-        KeyValue::new("B", "2"),
-        KeyValue::new("C", "3"),
-    ]
-});
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
@@ -88,39 +79,63 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
         result.err()
     );
 
-    // Opentelemetry will not provide a global API to manage the logger provider. Application users must manage the lifecycle of the logger provider on their own. Dropping logger providers will disable log emitting.
+    let meter_provider = result.unwrap();
+
+    // Opentelemetry will not provide a global API to manage the logger
+    // provider. Application users must manage the lifecycle of the logger
+    // provider on their own. Dropping logger providers will disable log
+    // emitting.
     let logger_provider = init_logs().unwrap();
 
-    let tracer = global::tracer("ex.com/basic");
-    let meter = global::meter("ex.com/basic");
+    // Create a new OpenTelemetryLogBridge using the above LoggerProvider.
+    let otel_log_appender = OpenTelemetryLogBridge::new(&logger_provider);
+    log::set_boxed_logger(Box::new(otel_log_appender)).unwrap();
+    log::set_max_level(Level::Info.to_level_filter());
 
-    let layer = OpenTelemetryTracingBridge::new(&logger_provider);
-    tracing_subscriber::registry().with(layer).init();
+    let common_scope_attributes = vec![KeyValue::new("scope-key", "scope-value")];
+    let tracer = global::tracer_provider()
+        .tracer_builder("basic")
+        .with_attributes(common_scope_attributes.clone())
+        .build();
+    let meter = global::meter_with_version(
+        "basic",
+        Some("v1.0"),
+        Some("schema_url"),
+        Some(common_scope_attributes.clone()),
+    );
 
-    tracer.in_span("operation", |cx| {
+    let counter = meter
+        .u64_counter("test_counter")
+        .with_description("a simple counter for demo purposes.")
+        .with_unit(Unit::new("my_unit"))
+        .init();
+    for _ in 0..10 {
+        counter.add(1, &[KeyValue::new("test_key", "test_value")]);
+    }
+    counter.add(1, &[KeyValue::new("test_key", "test_value")]);
+
+    tracer.in_span("Main operation", |cx| {
         let span = cx.span();
         span.add_event(
             "Nice operation!".to_string(),
             vec![Key::new("bogons").i64(100)],
         );
-        span.set_attribute(KeyValue::new(ANOTHER_KEY, "yes"));
+        span.set_attribute(KeyValue::new("another.key", "yes"));
+
+        info!(target: "my-target", "hello from {}. My price is {}. I am also inside a Span!", "banana", 2.99);
 
         tracer.in_span("Sub operation...", |cx| {
             let span = cx.span();
-            span.set_attribute(KeyValue::new(LEMONS_KEY, "five"));
-
+            span.set_attribute(KeyValue::new("another.key", "yes"));
             span.add_event("Sub span event", vec![]);
         });
-        info!(target: "my-target", "hello from {}. My price is {}. I am also inside a Span!", "banana", 2.99);
     });
+
     info!(target: "my-target", "hello from {}. My price is {}", "apple", 1.99);
 
-    let histogram = meter.f64_histogram("ex.com.two").init();
-    histogram.record(5.5, COMMON_ATTRIBUTES.as_ref());
-
     global::shutdown_tracer_provider();
-    logger_provider.shutdown();
-    global::shutdown_meter_provider();
+    logger_provider.shutdown()?;
+    meter_provider.shutdown()?;
 
     Ok(())
 }