+++ scripts/test.sh
@@ -11,5 +11,4 @@ cargo test --manifest-path=opentelemetry/Cargo.toml --no-default-features
 cargo test --manifest-path=opentelemetry/Cargo.toml --all-features -- --ignored --test-threads=1
 
 cargo test --manifest-path=opentelemetry/Cargo.toml --all-features
-cargo test --manifest-path=opentelemetry-jaeger/Cargo.toml --all-features -- --test-threads=1
 cargo test --manifest-path=opentelemetry-zipkin/Cargo.toml --all-features