+++ opentelemetry-stdout/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 
+- TraceExporter fixed to print InstrumentationScope's attributes.
 - Fix StatusCode in stdout exporter [#1454](https://github.com/open-telemetry/opentelemetry-rust/pull/1454)
 - Add missing event timestamps [#1391](https://github.com/open-telemetry/opentelemetry-rust/pull/1391)
 - Adjusted `chrono` features to reduce number of transitive dependencies. [#1569](https://github.com/open-telemetry/opentelemetry-rust/pull/1569)